+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `startsWith` and `endsWith` for `List`.
 - Added expanded set up apple targets for kotlin native.
 
+### Fixed
+- Fixed behavior of `isEqualToIgnoringGivenProperties` to only check accessible properties.
+
 ## [0.24] 2021-05-05
 
 ### Fixed